task_management/requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.1.7
django-filter==23.3
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
PyJWT==2.10.1
pytest==7.4.3
pytest-django==4.5.2
python-dotenv==1.0.0
redis==5.2.1
sqlparse==0.5.3
tzdata==2025.1